SNOW-2176180: code refactor 2
